Refactored SoftLayerUtils to ProductPackagePredicates
